+++ tests/Basket/BasketServiceTest.php
@@ -152,16 +152,17 @@ final class BasketServiceTest extends ApiTestCase
             $selectedPayment,
             true,
             $redirectUrl,
-            $cssUrl
+            $cssUrl,
+            'EN'
         );
 
         // @TODO : check that the two following values are normal
         $this->assertSame('', $paymentInformation->getHtml());
-        $this->assertNull($paymentInformation->getRedirectUrl());
+        $this->assertSame($redirectUrl, (string) $paymentInformation->getRedirectUrl());
 
         $orders = $paymentInformation->getOrders();
         $this->assertCount(1, $orders);
-        $this->assertSame(15, $paymentInformation->getParentOrderId());
+        $this->assertSame(43, $paymentInformation->getParentOrderId());
 
         $order = $orderService->getOrder($orders[0]->getId());
         $this->assertSame($orders[0]->getId(), $order->getId());
@@ -677,7 +678,14 @@ final class BasketServiceTest extends ApiTestCase
         $basket = $service->createEmptyBasket();
 
         $this->expectException(BasketIsEmpty::class);
-        $service->checkout($basket->getId(), 1, true, 'https://demo.loc/order/confirm');
+        $service->checkout(
+            $basket->getId(),
+            1,
+            true,
+            'https://demo.loc/order/confirm',
+            'https://demo.loc/custom.css',
+            'EN'
+        );
     }
 
     public function testGetTotalMarketplaceDiscount(): void